+++ test/deployer/Deployer.t.sol
@@ -152,7 +152,7 @@ contract DeployerTest is Test {
         deployer.setDescription(testFeature, "nothing to see here");
         deployer.authorize(testFeature, address(this), uint40(block.timestamp + 1 days));
 
-        vm.expectRevert(abi.encodeWithSignature("NoToken(uint256)", testTokenId));
+        vm.expectRevert(abi.encodeWithSignature("ERC721NonexistentToken(uint256)", testTokenId));
         deployer.ownerOf(testTokenId);
 
         (address instance, Nonce nonce) = deployer.deploy(testFeature, type(Dummy).creationCode);
@@ -165,7 +165,7 @@ contract DeployerTest is Test {
         vm.expectEmit(true, true, true, false, address(deployer));
         emit IDeployer.Removed(testFeature, Nonce.wrap(1), instance);
         assertTrue(deployer.remove(testFeature, nonce));
-        vm.expectRevert(abi.encodeWithSignature("NoToken(uint256)", testTokenId));
+        vm.expectRevert(abi.encodeWithSignature("ERC721NonexistentToken(uint256)", testTokenId));
         deployer.ownerOf(testTokenId);
 
         (instance, nonce) = deployer.deploy(testFeature, type(Dummy).creationCode);
@@ -202,7 +202,7 @@ contract DeployerTest is Test {
         Vm.Log[] memory entries = vm.getRecordedLogs();
         assertEq(entries.length, 1);
 
-        vm.expectRevert(abi.encodeWithSignature("NoToken(uint256)", testTokenId));
+        vm.expectRevert(abi.encodeWithSignature("ERC721NonexistentToken(uint256)", testTokenId));
         deployer.ownerOf(testTokenId);
 
         for (Nonce i = zero.incr(); nonce > i; i = i.incr()) {
@@ -236,6 +236,50 @@ contract DeployerTest is Test {
         vm.expectEmit(true, true, true, false, address(deployer));
         emit IDeployer.Removed(testFeature, nonce, instance);
         deployer.remove(testFeature, nonce);
+    }
+
+    function testNext() public {
+        deployer.setDescription(testFeature, "nothing to see here");
+        deployer.authorize(testFeature, address(this), uint40(block.timestamp + 1 days));
+
+        address next = Create3.predict(_salt(Feature.unwrap(testFeature), 1), address(deployer));
+        assertEq(deployer.next(testFeature), next);
+
+        (address instance, ) = deployer.deploy(testFeature, type(Dummy).creationCode);
+        assertEq(instance, next);
+
+        assertEq(deployer.ownerOf(Feature.unwrap(testFeature)), next);
+    }
+
+    function testPrev() public {
+        deployer.setDescription(testFeature, "nothing to see here");
+        deployer.authorize(testFeature, address(this), uint40(block.timestamp + 1 days));
+
+        (address firstInstance, Nonce firstNonce) = deployer.deploy(testFeature, type(Dummy).creationCode);
+        (address secondInstance, Nonce secondNonce) = deployer.deploy(testFeature, type(Dummy).creationCode);
+        (address thirdInstance, Nonce thirdNonce) = deployer.deploy(testFeature, type(Dummy).creationCode);
+        (, Nonce fourthNonce) = deployer.deploy(testFeature, type(Dummy).creationCode);
+
+        address prev = Create3.predict(_salt(Feature.unwrap(testFeature), Nonce.unwrap(thirdNonce)), address(deployer));
+
+        assertEq(prev, thirdInstance);
+        assertEq(deployer.prev(testFeature), prev);
+
+        deployer.remove(testFeature, fourthNonce);
+
+        assertEq(deployer.ownerOf(Feature.unwrap(testFeature)), prev);
+        assertEq(deployer.prev(testFeature), secondInstance);
+
+        deployer.remove(testFeature, secondNonce);
+
+        assertEq(deployer.ownerOf(Feature.unwrap(testFeature)), prev);
+        assertEq(deployer.prev(testFeature), firstInstance);
+
+        deployer.remove(testFeature, firstNonce);
+
+        assertEq(deployer.ownerOf(Feature.unwrap(testFeature)), prev);
+        vm.expectRevert(abi.encodeWithSignature("NoInstance()"));
+        deployer.prev(testFeature);
     }
 
     function testTokenURI() public {